refactor(components) : move ChannelAppointment form logic to ChannelAppointmentForm component and create other forms templates as well
