+++ frontend/app/channeling/upload-lab-report/page.tsx
@@ -1,8 +1,16 @@
+"use client";
 
-export default function UploadLabReport() {
+import { ChannelAppointmentForm } from "@/components/ui/forms/ChannelAppointmentForm";
+
+export default function UploadLabRepor() {
   return (
-    <div>
-      This is the patient&apos;s lab report uploading form
+    <div className="min-h-screen bg-gray-100 px-4 py-8">
+      <h1 className="text-black text-3xl font-bold text-center mb-6">
+        This is page is for uploading patient LAB reports
+      </h1>
+      <ChannelAppointmentForm />
     </div>
-  )
+  );
 }
+
+
