+++ frontend/app/channeling/view-my-appointments/page.tsx
@@ -6,3 +6,4 @@ export default function ViewAppointment() {
     </div>
   )
 }
+
