+++ frontend/components/ui/forms/index.ts
@@ -1,1 +1,3 @@
 // Note: add your custom ui elements in the above folder 
+
+export { ChannelAppointmentForm } from "./ChannelAppointmentForm";
